Move appendArc to dedicated arc.js module (SoC: DOM helper, not geometry)
